+++ app/Http/Controllers/SellerController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Seller;
 use App\Models\User;
 use Exception;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Illuminate\Validation\ValidationException;
@@ -20,8 +21,9 @@ class SellerController extends Controller
     public function create(Request $request)
     {
         try {
-            $adm = auth()->user();
+            $this->authorize('create', Seller::class);
 
+            $adm = auth()->user();
             $request->merge(['created_by' => $adm->id]);
             $validatedData = $request->validate($this->rules);
 
@@ -29,8 +31,9 @@ class SellerController extends Controller
             return response(['seller' => $createdSeller], Response::HTTP_CREATED);
         } catch (ValidationException $e) {
             return response()->noContent(Response::HTTP_BAD_REQUEST);
+        } catch (AuthorizationException $e) {
+            return response()->noContent(Response::HTTP_FORBIDDEN);
         } catch (Exception $e) {
-            var_dump($e->getMessage());
             return response()->noContent(Response::HTTP_INTERNAL_SERVER_ERROR);
         }
     }
@@ -55,6 +58,8 @@ class SellerController extends Controller
     public function update(Request $request, Seller $seller)
     {
         try {
+            $this->authorize('update', Seller::class);
+            
             unset($this->rules['created_by']);
             $validatedData = $request->validate($this->rules);
             $seller->update($validatedData);
@@ -62,19 +67,27 @@ class SellerController extends Controller
             return response(['seller' => $seller], Response::HTTP_OK);
         } catch (ValidationException $e) {
             return response()->json(['errors' => $e->errors()], Response::HTTP_BAD_REQUEST);
+        } catch (AuthorizationException $e) {
+            return response()->noContent(Response::HTTP_FORBIDDEN);
         } catch (Exception $e) {
-            return response()->json(['error' => 'Erro interno'], Response::HTTP_INTERNAL_SERVER_ERROR);
+            return response()->noContent(Response::HTTP_INTERNAL_SERVER_ERROR);
         }
     }
 
     public function destroy(Seller $seller)
     {
-        $adm = auth()->user();
-        $seller->deleted_by = $adm->id;
+        try {
+            $this->authorize('delete', Seller::class);
 
-        $seller->save();
-        $seller->delete();
-
-        return response()->noContent(Response::HTTP_OK);
+            $adm = auth()->user();
+            $seller->deleted_by = $adm->id;
+            $seller->save();
+            $seller->delete();
+            return response()->noContent(Response::HTTP_OK);
+        } catch (AuthorizationException $e) {
+            return response()->noContent(Response::HTTP_BAD_REQUEST);
+        } catch (Exception $e) {
+            return response()->noContent(Response::HTTP_INTERNAL_SERVER_ERROR);
+        }
     }
 }
